Change distributed directory structure
