+++ rl_algorithms/dqn/learner.py
@@ -1,4 +1,4 @@
-"""Learner for DQN Agent
+"""Learner for DQN Agent.
 
 - Author: Chris Yoon
 - Contact: chris.yoon@medipixel.io
